change detector_server to return the prediction value instead of a value after the threshold check so that the threshold is checked by the client

+++ detector-client/src/main/java/com/amaxilatis/metis/detector/client/dto/DetectionsDTO.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class DetectionsDTO {
     private int w;
     private int h;
-    private int[][] predictions;
+    private double[][] predictions;
 }
